making sure tests have right state before testing (#2253)
* making sure tests have right state before testing

+++ playwright/e2eTests/altinn3/klientadministrasjon/klientadministrasjon.spec.ts
@@ -54,6 +54,15 @@ test.describe('klientadministrasjon', () => {
     const actor = { pid: '16830197862', org: '311164651', orgName: 'KLARTENKT UPRESIS ISBJØRN SA' };
     const agent = { pid: '30877795760', name: 'TROVERDIG JUICE' };
 
+    test.beforeEach(async () => {
+      try {
+        await api.deleteClientDelegationAgent(actor.pid, actor.org, agent.pid);
+      } catch {
+        /* ignore if nothing to clean */
+      }
+      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
+    });
+
     test('slett bruker', async ({
       page,
       accessManagementFrontPage,
@@ -61,8 +70,6 @@ test.describe('klientadministrasjon', () => {
       login,
       aktorvalgHeader,
     }) => {
-      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
-
       await test.step('Logg inn', async () => {
         await page.goto(env('BASE_URL'));
         await login.LoginToAccessManagement(actor.pid);
@@ -86,7 +93,9 @@ test.describe('klientadministrasjon', () => {
       });
     });
 
-    test.afterEach(async () => {
+    test.afterEach(async ({}, testInfo) => {
+      if (testInfo.status === 'passed') return;
+
       try {
         await api.deleteClientDelegationAgent(actor.pid, actor.org, agent.pid);
       } catch (error) {
@@ -104,6 +113,24 @@ test.describe('klientadministrasjon', () => {
     };
     const agent = { pid: '28822449737', name: 'SINDIG BRUD' };
 
+    test.beforeEach(async () => {
+      try {
+        await api.deleteConnection(client.connectionPid, client.org, [actor.org]);
+      } catch {
+        /* ignore if nothing to clean */
+      }
+      try {
+        await api.deleteClientDelegationAgent(actor.pid, actor.org, agent.pid);
+      } catch {
+        /* ignore if nothing to clean */
+      }
+
+      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
+        posttjenester,
+      ]);
+      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
+    });
+
     test('deleger klient til en bruker fra klientfanen', async ({
       page,
       accessManagementFrontPage,
@@ -111,11 +138,6 @@ test.describe('klientadministrasjon', () => {
       login,
       aktorvalgHeader,
     }) => {
-      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
-        posttjenester,
-      ]);
-      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
-
       await test.step('Logg inn', async () => {
         await page.goto(env('BASE_URL'));
         await login.LoginToAccessManagement(actor.pid);
@@ -175,6 +197,24 @@ test.describe('klientadministrasjon', () => {
     };
     const agent = { pid: '47822800420', name: 'UVANLIG FREDAG' };
 
+    test.beforeEach(async () => {
+      try {
+        await api.deleteConnection(client.connectionPid, client.org, [actor.org]);
+      } catch {
+        /* ignore if nothing to clean */
+      }
+      try {
+        await api.deleteClientDelegationAgent(actor.pid, actor.org, agent.pid);
+      } catch {
+        /* ignore if nothing to clean */
+      }
+
+      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
+        posttjenester,
+      ]);
+      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
+    });
+
     test('deleger klient til en bruker fra brukerfanen', async ({
       page,
       accessManagementFrontPage,
@@ -182,11 +222,6 @@ test.describe('klientadministrasjon', () => {
       login,
       aktorvalgHeader,
     }) => {
-      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
-        posttjenester,
-      ]);
-      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
-
       await test.step('Logg inn', async () => {
         await page.goto(env('BASE_URL'));
         await login.LoginToAccessManagement(actor.pid);
@@ -241,6 +276,14 @@ test.describe('klientadministrasjon', () => {
     };
     const agent = { pid: '04905999376', name: 'FORSTÅELSESFULL TRAFIKKORK' };
 
+    test.beforeEach(async () => {
+      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
+        posttjenester,
+      ]);
+      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
+      await api.delegerKlientTilBruker(actor.pid, actor.org, client.org, agent.pid, posttjenester);
+    });
+
     test('Slett fullmakt for en bruker fra brukerfanen', async ({
       page,
       accessManagementFrontPage,
@@ -248,12 +291,6 @@ test.describe('klientadministrasjon', () => {
       login,
       aktorvalgHeader,
     }) => {
-      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
-        posttjenester,
-      ]);
-      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
-      await api.delegerKlientTilBruker(actor.pid, actor.org, client.org, agent.pid, posttjenester);
-
       await test.step('Logg inn', async () => {
         await page.goto(env('BASE_URL'));
         await login.LoginToAccessManagement(actor.pid);
@@ -305,6 +342,14 @@ test.describe('klientadministrasjon', () => {
     };
     const agent = { pid: '03924296991', name: 'ORANSJE TEST' };
 
+    test.beforeEach(async () => {
+      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
+        posttjenester,
+      ]);
+      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
+      await api.delegerKlientTilBruker(actor.pid, actor.org, client.org, agent.pid, posttjenester);
+    });
+
     test('Slett fullmakt for en bruker fra klientfanen', async ({
       page,
       accessManagementFrontPage,
@@ -312,12 +357,6 @@ test.describe('klientadministrasjon', () => {
       login,
       aktorvalgHeader,
     }) => {
-      await api.addConnectionAndPackagesToUser(client.connectionPid, client.org, actor.org, [
-        posttjenester,
-      ]);
-      await api.addClientDelegationAgent(actor.pid, actor.org, agent.pid);
-      await api.delegerKlientTilBruker(actor.pid, actor.org, client.org, agent.pid, posttjenester);
-
       await test.step('Logg inn', async () => {
         await page.goto(env('BASE_URL'));
         await login.LoginToAccessManagement(actor.pid);
